calendar/requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
icalendar>=5.0.0
caldav>=0.9.0
requests>=2.28.0
